# docs/api/User_Api_Swagger.yaml
...
openapi: 3.0.0
info:
  title: ESP RainMaker Neo Auth & Accounts API
  version: 0.8.0
  description: |
    Sign-in and account management for the ESP RainMaker Neo cloud.
    Anything that produces or rotates an access token lives here; once
    you have a token, the rest of the platform lives in the
    **ESP RainMaker Neo Platform API**

    ESP User is the platform's own OAuth 2.1 / OIDC issuer: every token a client
    holds is minted here, regardless of how the user proved their identity.
    Admins authenticate directly against their AWS Cognito user pool.

    Login methods (all end in OUR tokens; upstream credentials never reach clients):

    - **Federated login (default)**: `GET /oauth2/authorize` starts a browser
      authorization-code + PKCE flow. With a single enabled identity provider the
      user is auto-redirected to it; the
      broker exchanges the upstream code server-side and issues OUR authorization
      code at the client's `redirect_uri`.
    - **Native password APIs** (`/v1/user/auth/*`): the main-era
      native surface. Credentials are verified against the bundled Cognito pool server-side and converted to our tokens.

    Token lifecycle after login:

    1. `POST /oauth2/token` (`grant_type=refresh_token`) - rotate the token set
       (native `/v1/user/auth/token/refresh` is reuse-tolerant instead).
    2. Send `access_token` as a `Bearer` `Authorization` header on subsequent
...
       must come from the same sign-in.
...
  contact:
...
  license:
...

...
servers:
...
paths:
...
  /v1/user/auth/token:
    post:
      summary: Password sign-in
      description: |
        Signs in with `username` and `password`, verified against the bundled Cognito pool, and returns an ESP User token set.
      tags:
        - User Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyTokenRequest"
      responses:
        "200":
          description: ESP User token set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenSet"
        "400":
          description: "`Invalid request body`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: |
            `Authentication failed` — wrong password, or no such account.
            `Signin failed. Account not verified — reset your password.` — correct password, account never verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/user/auth/token/refresh:
    post:
      summary: Refresh tokens
      description: |
        Exchanges a refresh token for a new token set. The refresh token rotates: store the new one, because replaying a spent one ends the login.
      tags:
        - User Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyRefreshRequest"
      responses:
        "200":
          description: ESP User token set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenSet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/user/auth/signup:
    post:
      summary: Create an account
      description: |
        Creates an account with exactly one of `email` or `phone_number` plus a `password`, and sends a verification code to confirm at `signup/verify`.
      tags:
        - User Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacySignupRequest"
      responses:
        "201":
          description: Account created; verification required.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  requires_verification:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example:
                      Code sent. Existing RainMaker users must signin or reset
                      password.
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/user/auth/signup/verify:
    post:
      summary: Confirm signup
      tags:
        - User Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                email:
                  type: string
                phone_number:
                  type: string
                code:
                  type: string
                  description: Verification code Cognito delivered at signup.
      responses:
        "200":
          description: "`Verified successfully. You can now login.`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: |
            `Invalid code or RainMaker account already exists. Try signin or reset password.` — one answer for every failure: wrong, expired or empty code, an account that does not exist, and an account that is already confirmed. The product name appears only when the resolved provider is named RainMaker.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/user/auth/password-recovery:
    post:
      summary: Start password recovery
      tags:
        - User Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  description: The account email or phone number.
      responses:
        "200":
          description: Reset code sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: If your account exists, you will receive a code
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/user/auth/password-recovery/confirmation:
    post:
      summary: Confirm password recovery
      tags:
        - User Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, code, new_password]
              properties:
                username:
                  type: string
                code:
                  type: string
                new_password:
                  type: string
      responses:
        "200":
          description: Password reset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/user/auth/password:
    post:
      summary: Change password
      description: |
        Changes the caller's password. Requires this issuer's access token and the current password. Other sessions stay signed in.
      tags:
        - User Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [access_token, old_password, new_password]
              properties:
                access_token:
                  type: string
                  description: Access token issued by this service.
                old_password:
                  type: string
                new_password:
                  type: string
      responses:
        "200":
          description: Password changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Missing field.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/user/auth/signout:
    post:
      summary: Sign out
      description: |
        Ends the login the refresh token belongs to (its refresh family), the same as `/oauth2/revoke`. `global: "true"` is refused with `400`; sign out each login with its own refresh token.
      tags:
        - User Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                global:
                  type: string
                  description:
                    Accepted for compatibility; `"true"` is rejected with `400`.
      responses:
        "200":
          description: Signed out.
        "400":
          description:
            Refresh token missing, or a global signout was requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  schemas:
...
    LegacyTokenRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: The account email or phone number.
        password:
          type: string
          format: password
    LegacyRefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Refresh token from this surface; spent once redeemed.
    LegacySignupRequest:
      type: object
      required: [password]
      properties:
        email:
          type: string
          description: Exactly one of email or phone_number.
        phone_number:
          type: string
          description: E.164; exactly one of email or phone_number.
        password:
          type: string
          format: password
    TokenSet:
      type: object
      description: The issued ESP User token set.
      properties:
        access_token:
          type: string
          description: The access token.
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: integer
          example: 3600
        refresh_token:
          type: string
          description: Refresh token; rotated on each use.
        id_token:
          type: string
          description: Present only if `openid` was requested.
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Detailed error message
          example: "Invalid request"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Unauthorized"
...
  securitySchemes:
    CognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      description: |
        Admin Cognito User Pool JWT, for the admin APIs (`/v1/admin/*`).
        End-user routes take an ESP User token instead — see `bearerAuth`.
      x-amazon-apigateway-authtype: cognito_user_pools
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An ESP User access token, sent as a Bearer token. The token is verified
        against the provider's published public keys.
    clientBasicAuth:
      type: http
      scheme: basic
      description: |
        OAuth client authentication (RFC 6749 §2.3.1) as HTTP Basic
        `base64(client_id:client_secret)`. A confidential client sends its secret;
        a public client sends its client_id with an empty secret.
security: []